# spring-server/target/classes/swagger_GamificationPlatform.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Gamification
  description: An API to implement gamification features in the swagger-2.0 specification. 
    
    ToDo 
      (- /rules)
    
  contact:
    name: Guillaume Serneels
    email: guillaume.serneels@heig-vd.ch
    url: www.heig-vd.ch
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: localhost
basePath: /api
schemes:
  - http
paths:
...
  /pointScales:
    get:
      summary: Returns every pointScales
      description: |
        Returns every pointScales
      operationId: findPointScales
      produces:
        - application/json
      parameters:
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: pointScales response
          schema:
            type: array
            items:
              $ref: '#/definitions/pointScale'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    post:
      summary: Creates a new pointScale
      description: |
        Creates a new pointScale
      operationId: addPointScale
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: pointScale
          in: body
          description: pointScale object to add to the store
          required: true
          schema:
            $ref: '#/definitions/pointScale'
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: pointScale response
          schema:
            $ref: '#/definitions/pointScale'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /pointScales/{pointScaleId}: 
    get:
      summary: Returns a single pointScale
      description: |
        Returns a single pointScale
      operationId: findPointScale
      produces:
        - application/json
      parameters:
        - name: pointScaleId
          in: path
          description: ID of pointScale to fetch
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: pointScale response
          schema:
            $ref: '#/definitions/pointScale'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    put:
      summary: Update an existing pointScale
      description: |
        Update an existing pointScale
      operationId: UpdatePointScale
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: pointScale
          in: body
          description: pointScale object to add to the store
          required: true
          schema:
            $ref: '#/definitions/pointScale'
        - name: pointScaleId
          in: path
          description: Id of the pointScale that needs to be updated
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: pointScale response
          schema:
            $ref: '#/definitions/pointScale'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    delete:
      summary: Delete an existing pointScale
      description: |
        Delete an existing pointScale
      operationId: DeletePointScale
      produces:
        - application/json
      parameters:
        - name: pointScaleId
          in: path
          description: Id of the pointScale that needs to be deleted
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '204':
          description: pointScale deleted successfully
        '404':
          description: pointScale not found
        '400':
          description: Invalid Id supplied
...
  /users:
    get:
      summary: Returns every users
      description: |
        Returns every users
      operationId: findUsers
      produces:
        - application/json
      parameters:
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: users response
          schema:
            type: array
            items:
              $ref: '#/definitions/user'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
...
  /users/{userId}/badges:
    get:
      summary: Returns a user's badges
      description: |
        Returns a user's badges
      operationId: findUserBadges
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: badges response
          schema:
            type: array
            items:
              $ref: '#/definitions/badge'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /users/{userId}/pointScales:
    get:
      summary: Returns a user's pointScales
      description: |
        Returns a user's pointScales
      operationId: findUserPointScales
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          type: integer
          format: int64
        - name: token
          in: header
          description: token to be passed as a header
          required: true
          type: string
      responses:
        '200':
          description: pointScales response
          schema:
            type: array
            items:
              $ref: '#/definitions/pointScale'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
definitions:
  badge:
    type: object
    required:
      - id
      - name
      - app
      - userId
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      app:
        type: integer
        format: int64
      tag:
        type: string
      userId:
        type: integer
        format: int64
  pointScale:
    type: object
    required:
      - id 
      - name
      - app
      - userId
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      app:
        type: integer
        format: int64
      description:
        type: string
      token:
        type: string
      userId:
        type: integer
        format: int64
  user:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
...
  errorModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string